Addressed issue #7 - .bombing of other bots
Other bots listed in the config file will now be ignored by .bomb,
as will the bot itself.
I'm not entirely sure where bug #8 is coming from, as the input and color
are both .lower()'ed. Tryed adding .rstrip(' '), in case that's the problem,
but it's still not quite right. Also still does not reject non-existant colors.