'\', alt separators, removed meant, added /i
The find module now can accept literal \ characters in both the orig and replacement of s///.

Any non-word character can be used as a separator, e.g. 's$foo$bar$'

The meant function was removed temporarily. It will be either reinstated or incorporated into the findreplace function.

Flag parsing now works differently. Any word characters (by unicode standards) following the final separator are considered to be flags. Unrecognized ones are ignored. For example, 's/foo/bar/kg' would do a global replace, but ignore the k. 's/foo/bar/ kg' would *not* do a global replace. The 'kg' is considered to be a comment because of the whitespace.

Whitespace parsing works differently. Whitespace is allowed before and after the 's'. 's/foo/bar/' is the same as ' s/foo/bar' is the same as 's /foo/bar/'.

The 'i' flag has been added for case-insensitive matching.