Separator must be non-word by *unicode* standards.
Previously, it was possible to use characters like á as a separator, but that makes no sense (someone could type a sentence like "sí, a mí me gusta esta música" and it would be interpreted like 's/, a m/ me gusta esta música' which was probably not intended). Don't worry, there are still thousands of valid separators, like '˘', '¥', '¤', etc.