Move bitly initialization to a setup() procedure
Some modules, for examples rss, imports url.py. This seems to me a bit ugly bit we didn't say we don't support this, so this change will supress the bitly warning the second time url.py is ran.

That said, bitly configuration should move to jenni.config, or maybe just be removed, I don't see what's the use for it