[weather] Overhaul to use Yahoo weather.
Existing weather location settings will need to be redone, as WOEID can't recognize ICAO codes, even a little. The icao column in the settings table is no longer used; a new woeid column was added. It's sad to be moving away from the public domain NOAA info, but I'm not sure what the terms are on the site we were using to get coordinates to match with our ICAO coordinates (which were out of date anyway).