[core] Properly consider final argument preceded by :
Previously, such arguments would be assumed to be text, when in fact
the RFC specifies that they are to be treated equally to any other
argument. This caused problems on servers running UnrealIRCd, because
it precedes the channel in a JOIN command with a colon. As such, it
would be properly seen as trigger.sender on some other networks, but
not others.