[core] Add systemd service file and other default config files for distros
See contrib/README for more information.
Also modified willie.py to support a new config key in [core]:
not_configured will cause the bot to refuse running if set to true.
This is useful to force users to configure the bot before trying to run it as the config wizard is not ideal for willie running as a service.

Issue #262