[rss] Convert old table format to new
If the new format table doesn't exist on setup, create it, check for the
old format table, and add any feeds and articles into the new one.

The old tables can then be dropped, but I've commented this out because
it makes it impossible for
any other module to use a table named "rss" or "recent". I'd suggest
standardizing DB table names in modules to be prepended with the module
name, so this isn't an issue.