[rss] Numerous tweaks to improve usability and prevent spamminess
Don't debug to everywhere for anything.
Enable checking feeds by default.
Don't stop checking when there are no feeds. It's worth the extra
database call every five minutes to not require rss to be restarted when
the first feed is added. Perhaps, in the future, we could automatically
scale down the period when there are no feeds or something.

Also, PEP8ify - issue #125