[core, modules] Python 3 support!
Issue #236

I'm happy to announce that with this commit, all basic Willie functionality
can be ran on Python 3!

There are few bugs remaining to iron out and few modules which fail to load
and need to be fixed, but all the basic functionality now works well with
Python3.

No extra dependencies introduced, Python 2.7 is obviously still supported.