[core] fix pedantic newlines in error()
PEP8 was invented to make code MORE readable, not less. This insane amount of
line breaking makes this code much less readable, hence it's better to go over
the 80 char limit a bit.

Anyone who changes these back to the previous formatting will be cursed by the
gods of code-readability.