.movieposter: print movie posters as ASCII art
This change adds a .movieposter command to the movie module to retrieve
and print movie posters as ASCII art using Python's aalib and Image
modules if they are available.

TODO: need to add support for configuring screen width/height, and use
of AnsiScreen (works if your IRC client is running in an ANSI-compatible
terminal).