[chanlogs] Security fix: properly handle PART messages
This is kindof a security fix. Willie created files with the PART messages
as file names. I thought it would be a directory traversal vlun
when I first noticed this behaviour, but fortunately it isn't because
Willie would add "" around the first part of the filename.
yay for safe formatting.

Anyway, the reason for this bug was somone using `trigger` instead of
`trigger.sender`