allow op command for bot admins everywhere
this is more consistent with the command's documentation, which does
say that the bot will op the user that gave the command if no user is
specified. yet this would never work if the user is not op.

i think admins will expect a bunch more commands to work (like voice,
deop and so on) so maybe a more unified access control system would be
necessary, but that should be a good start.