try to find a new nick if previous failed
we really need to have a config option for the max here, but i want to
test the waters before going any further at fixing #300, since it was
proposed to only use GHOST.

those approaches should be complementary: if we fail at using GHOST
with nickserv, we should fallback to this dumb routine.

choosing a nick should probably be encapsulated in a function instead
of just being thrown in the terminator routine. i'm also unsure about
the variables names and would welcome feedback on that.