core: Fix starting willie when an empty pidfile exists
It seems that some of the pidfile logic got messed up, spewing a "Willie
is not running" error when neither --quit nor --kill was specified. This
simplifies the elif to only check for --quit or --kill. The preceding if
already checks if the pidfile contains a pid, which is also running, so
when we get to this elif we already know that the pidfile is empty, or
contains a pid that is no longer running.