core: Only replace . with the prefix in examples
Previously, the first character of an example was always forced to be
the current prefix. This might lead to unexpected results if the first
character isn't intended to be the prefix. Since the @example decorator
should normally be called with an example that uses the default . as the
prefix, this changes the processing to only replace the first character
of an example if it is actually a .

Additionally, the replacing is slightly modified, so it now properly
works when the new prefix is not a single character.

Finally, the @example decorator documentation is updated to indicate
that this replacement is applied.