config: Move default for log_raw option into Config class
Previously, this default was defined in irc.py, whereas all other
defaults are in the Config class.

Additionally, this sets the default value in the actual ConfigParser
object, instead of in the parsed ConfigSection object. This means we
must use 'true' instead of 'True', since all ConfigParser values are
strings.