Clear job queue when reloading
Just a stop-gap to make #831 not a regression so 6.0 can get out the
door. It's not the right solution, but it's slightly better than what
was there before. Only reloading a module with a job actually in it will
kill all the jobs now, rather than any module. So, progress?