help: Split commands into groups by module
Close #662, close #791. They needed to be reworked to deal with the new
6.0 loader. Same general idea, so thanks to @qsantos and @firerogue for
their work there. Eventually, I'd like to open up that category to be
settable to something other than the module name, but I'm not sure how I
want to do that so we'll leave it as an undocumented little private hack
for now.