safety: Properly get the link from the input
Okay, so this is kind of a funny story. Turns out I used `trigger`
instead of trigger.group(1).

This basically mean the safety module has always been broken and nobody
ever noticed! If you posted a link with some text next to it, urlparse
would show the netloc as empty... which would make the safety module
search for an empty string in the blacklists (and submit an empty
string to Virus Total).

The only reason we noticed it is sometimes recently, a newline was added
in the malwaredomains DB file, which we didn't filter out when reading.

The result was that any link would be marked as evil,
because it's "on the list".

This resolves #870