Resolve #899
This is a bit of a hacky fix, the problem is that the obj at this point could be any part of the module file, including an import or docstring or whatever. This just checks if obj is callable and then unregisters in that case. If it's not callable, we just ignore it. (Should be safe enough to ignore)