Add 'UTC' as a timezone abbreviation for UTC timestamps
Currently, if no timezone is given, we use None. The default of datetime
in that case is to use a UTC timezone *without a name*. This is
confusing, as it's not clear from the printed timestamp, that it's in
UTC. This instead uses an actual implementation of UTC, with a name, as
a default.