Added key support for movie module
As of May 8, 2017, the OMDb API has gone private. From now on it requires a paid API key to function.
This change adds support for an api key in the configuration file.
It also changes the module command examples so that they will not fail the CI tests if there is no API key.