Allow any 2.x version of requests
In particular, the plan to require a newer version of PRAW will benefit
from this, since recent versions of PRAW prefer more recent versions of
`requests`. Testing with the latest `requests` (2.18.4 as of writing)
showed no issues with all the Sopel modules and functions that I was
able to try out.