search: Make tests less fragile
Any site that might redirect based on language headers is unsuitable for
use in tests. Apparently Wikipedia does that in some regions, or at
least DuckDuckGo does. Random test failures depending on the IP address
in use when tests are run are not acceptable, so remove Wikipedia from
search module tests.