loader, module: Fix help_prefix cleanup & example docs
Callable cleanup in loader now always uses help_prefix's default value
to replace. The documentation for module.example() reflects this, and I
also moved the constructor documentation up to where it would be
included in the auto-generated API docs (so the parameters are now
visible in the public documentation without reading the code).