Fix IPython requirement for real
Either pip isn't smart enough to figure out which Python versions a
package version is compatible with and pick one it can use, or IPython
is packaged in such a way that it breaks pip's smarts.

Whichever it is, Sopel will have to be the smart one and hold both their
hands to avoid breaking things.

Resolves #1308