weather: deduplicate code
The bulk of the old weather() and forecast() function bodies were
basically identical. This means we can just replace the commands with
wrapper functions and perform mostly the same steps with minimal
conditional branching in a single say_info() function.

Also took the opportunity to switch the API calls to HTTPS.