search: remove tests that hit Bing
Bing returning inconsistent results has been the number-one cause of CI
failures for the last several months. Basically the only cause, in fact.

Kicking the CI occasionally to re-run a failed job was tolerable, but
it's reached a point where seemingly no number of job restarts will get
a successful test run out of Bing, so away the tests will go.

Coveralls will probably yell at me for merging this, but I don't care.
I'm tired of seeing red X's on pull requests because of this.