config: add [core] 'interpolation' and 'interpolation_extended' params
RawConfigParser is the legacy python2 parser and does not allow/perform
interpolation.  More complex config files can benefit greatly from
interpolation and so we should use ConfigParser and allow interpolation.

The new params allow backwards compatibility if neither param is set
(or they are set to False).  If either basic or extended interpolation
is desired, either param can be added to the [core] section and set to
True to cause sopel to change its parser to the newer ConfigParser
and re-read the file, with either basic (the default) or extended
interpolation enabled.

Note that as extended interpolation is not available in Python2, if
the config file has 'interpolation_extended' enabled when running
Sopel under Python2, a RuntimeError is raised.