wiktionary: fix failed lookup for certain words
Fixes sopel-irc/sopel#1361 where certain words do not return
a definition when one is available.

Words that include references in definitions expose a missed condition
check in parsing that is not relevant for most cases. References use the
<sup> tag, which in reference has its own 'id="' attribute, causing
the `mode` variable in the loop to be reset, even though a definition
should be parsed from that line. These superscript references are also
not parsed out of the definition line.