logging: add format and datefmt to allow configuring the log output from cfg file
also add IRC channel-specific logging params to set level, format, and datefmt
for IRC channel logs, different from the local logs.  This can be useful when
the format of the IRC logs should be different than the format of the local logs.
This is also quite useful when the local log level is set to DEBUG, as logging
a huge volume of DEBUG logs to IRC is rarely useful.