run_script: replace global homedir
The `sopel.run_script.homedir` variable was a "global" variable that
belonged to the configuration realm. Therefore, I moved it to the
`sopel.config` module, and renamed it to `DEFAULT_HOMEDIR`.

This name:

* follow the PEP8 naming convention (not required but nice to have)
* move from a "run thing" module to a "config" module, which is more
  appropriate
* express the need for being the "default" location for configuration,
  since a local configuration can be used that might not be in it

All of this is unecessary, and more "nice to have", but it could help
fixing #1243 - if that makes sense.