run_script: replace sys.exit by return
When you install Sopel using `pip install` (or `setup.py develop`), it
installs a script, that contains this interesting bit:

    sys.exit(
        load_entry_point('sopel', 'console_scripts', 'sopel')()
    )

Wich is the same as:

    from sopel.run_script import main
    sys.exit(main())

So I replaced all `sys.exit(n)` in `run_script` by `return n` instead.

It works exactly the same, but it's easier to test or interact with
outside of the pre-defined command line tool. This is a step forward
helping other to reuse Sopel's code in their own way.