readme: make document title & move badges above intro
With the Open Collective badges added, there isn't enough space on the
"Introduction" heading line for all of the content in GitHub's render.
While it's trickier to preview the document on PyPI (since it pulls the
readme from uploaded packages, and we're not making new releases just to
test documentation), we certainly can make sure the readme looks OK on
GitHub, which is probably the place where most users see the project
first anyway.