wiktionary: improve control flow
Move all parts of speech to constant to make iteration easier; this
helps avoid the messy `if-elif-else` chain that results when parsing
for definitions, and makes it easier to change acceptable parts of
speech if/when Wiktionary changes specifications.

_Note: not all parts of speech are actually parts of speech, but they
are listed in the specification under the "Parts of speech" heading, so
just easier to call them that. See the comments in the_ `PARTS_OF_SPEECH`
_constant for details._

For now, this refactor does not break any old features, but also
does not yet correctly parse the newly added parts of speech
(e.g., phrases, proper nouns).