core: 2-bytes unicode characters are not truncated anymore
To send a long message, Sopel needs to split the message in multiple
lines of text, and used to split the bytestring version of the
unicode text.

This would cause an issue when 2-bytes unicode characters get truncated
in half, displaying unwanted message.

In this commit, I split the unicode string properly, and I truncate it
until its encoded version is below the max length.

See also the test/test_tools.py file for all the use-case handled by
this change.