Don't require pyenchant on Python 3.7 or above
pyenchant is broken on newer distributions, and has been unmaintained
for quite a while. Since only the spellcheck module needs it, and a
rewrite of that module to switch libraries is coming Soon™, we might as
well make Sopel installable in the mean time on systems that can't pull
down a working pyenchant build.

See #1452. Alternative to #1453.