Rip out the wiki's still-beating heart
This all started when I realized the /docs/ index page actually linked
to the wiki. Honestly, there weren't as many references to Sopel's wiki
as I thought there would be. But they were in some important places,
like the README file. Probably not good to have the README point at
documentation that we've essentially killed.

[Skip CI]