run_script: use variables for error codes
When the bot exits due to an error, its exit code must not be 0, so we
return a non-0 value. By default, it's 1, but to prevent systemd to
restart the bot in some specific error cases, we return 2 instead.

For that purpose, we use 2 variables:

* `ERR_CODE` (1), the default error code
* `ERR_CODE_NO_RESTART` (2), the error code to be returned when the
  program can not be restarted without a modification

This way, these values may change in the future (even if it is very
unlikely) without breaking the code, and more importantly, this
behavior is now easier to document (with docstrings and clear names).