etymology: quote URL parameters
The `ety` command now properly encodes the search term used in the URL
to request www.etymonline.com.

This way, the command `.ety coup d'etat` returns something like:

> "1640s, from French coup d'étate, literally 'stroke of the state'
> (see coup). […]" - https://www.etymonline.com/word/coup%20d%27etat

Instead of this:

> Can't find the etymology for "coup d'etat".
> Try https://www.etymonline.com/search?q=coup d'etat

Which was very unfortunate, because a) it didn't find the answer
and b) it gave an invalid URL to search for an answer.

It also encodes the search URL when it does not found a result:

    <me> .ety not found
    <Sopel> Can't find the etymology for "not found".
            Try https://www.etymonline.com/search?q=not%20found

So that the user can copy and paste the URL.