unicode: remove `u` prefix
And also extract the code into the `get_codepoint_name` function.

The `hex` function returns a `str` object, and you can not use a
bytestring object when you need a unicode one. Which is a problem,
because py2's str is py3's bytestring, and the py3's str is
py2's unicode.

The previous implementation used to convert everything into `str`
and to call it a day. It "works", but requires some hack, for instance:

    ord(u'' + arg)
    point.rjust(4, str('0')

Which can be quite confusing. To make things easier to read, the new
implementation defines `unicode` for py3 as an alias for `str`, and
uses it in the code once. From that line:

    unicode(hex(ord(char)))[2:]

Everything is working with unicode strings (which is more consistent)
instead of working with bytestrings in py2, and unicode in py3.

That being said, I'm amazed by the hack that was in place before,
because it shows how complex the situation trully is. Unicode is hard!