coretasks: Consistently update privilege tracking
Both `handle_names()` and `_record_who()` should update both
`bot.privileges` and `bot.channels`. The fact that they don't is a
leftover bug from who knows when.

We're also no longer happy with throwing an `assert` in released code,
so the consistency check in `track_modes()` is now a logger warning.