plugin: load plugin only once by name
Each plugin has a name, but it might not be unique. These plugins have
the same "name" (here, "xkcd"):

* `sopel.modules.xkcd`,
* `~/.sopel/modules/xkcd.py`,

but only one can be loaded as the `xkcd` plugin.

For that purpose, the previous implementation used a `dict`, without
consideration for loading order.

The new implementation uses an `OrderedDict`, to keep the same behavior
(last plugin found overrides previous ones), but also to make sure to
always load the core plugin at the end of the process.