config: fix `ListAttribute` wizard logic
Adding support for commas in `ListAttribute` broke the logic in the
`configure` method used during the wizard for prompting and processing
input.

Previously, because serialization was just basic concatenation, the
current items in the list could be represented as a serialized string.
That string would appear as the default value in the prompt, and it was
easy to simply append a new item to the string for serialization. After
supporting commas, this became an issue because converting the default
value to a string before taking new input would lead to the
`serialize`-er escaping the commas as if the default values were one
item. e.g.:

  - Old items: `['spam', 'eggs', 'sausage', 'bacon']`
  - Wizard would prompt: `Enter ... [spam,eggs,sausage,bacon]`
  - Accepting default values (and adding new values as needed) would
    append the string as one item to the list
  - 'serialize`-ing would result in a list like
    `['spam,eggs,sausage','bacon']` because the default item is
    `serialize`-d to become: `spam\,eggs\,sausage,bacon`

Now, the default prompt display is decoupled from the default value
handling logic. Additionally, the current implementation wraps the items
in the default values of the prompt with `"` to help delineate items
more precisely.