Start a new branch in which to draft NEWS entries
These are pending changes for Sopel 7, i.e. PRs that were merged to
master and should be mentioned in the changelog.

The whole branch should be squashed down into the version bump when
release time comes, rather than preserving all the commits from this one
(they'll be mostly one- or two-liners, so not worth keeping).

Maintaining this branch is merely a convenience, intended to eliminate
the arduous process of going through the commit log for Sopel 7.0.0,
which will probably contain hundreds of commits once all is said and
done, to write the changelog entry all at once. Doing that is
error-prone, while adding changes as they're merged will make sure
everything that needs to be mentioned, is. (Well, it will once I fill in
the PRs that have already been merged for 7.0.0.)