NEWS: Add issue/PR references to 7.0 entries
Might decide to link a different way (with support from the website
changelog-generator script), because honestly this method is really
tedious. I'm not a fan of the huge list of links, either.

Auto-linking to GitHub issue/PR numbers should be easy enough. The only
instances of `#\d+` in the file are GitHub references, in fact. Probably
quite simple to just insert a conversion step into document_versions.py
that converts anything that looks like an issue reference into a link.
The biggest challenge would be figuring out whether to use `/issues/` or
`/pull/` in the URL, but that's not critical—GitHub understands both no
matter what and redirects to the correct type.