fix: reload modules with `@rule`'d callables imported into __init__.py
If a callable was imported like from .module import *, the recursion
used in `reload.py` would catch the callable twice. The first time
around, when the object name has the format `module.submodule...`, the
callable obj has the rules as uncompiled regexp strings and no
associated priority, and so causes issues in the `bot.unregister`
method. The second time, the object name is just the object name and now
rules have been compiled (with associated priorities). This is **only**
observed with the `@rule`-decorated callables, and not with
`@command`-decorated ones.

While the object that with the `module.submodule...` in the name does
actually exist, it is not used anywhere and just hangs around. It's like
a human wisdom tooth... Since it has no effect on bot function, as it is
never registered and the bot is unaware of its existence, we can just
ignore it by checking if the callable object name contains a `.` Upon
finding a `.`, we just skip `unregister`'ing it.

Also, while dealing with this, I added a check to ensure that compiled
rules are added to a callable's rules list only once. While not directly
contributing to the issue, it was tangentially related, and worth
refactoring.