auth: split 'auth_method' into 'server_auth_method' and 'nick_auth_method'
currently only a single authentication method may be specified, so it is
not possible to do both server-based and nick-based authentication.  This
splits that into 'server_auth_method' which can be either 'server' and 'sasl'
and 'nick_auth_method' which can be 'nickserv', 'userserv', 'authserv',
or 'Q'.  This allows, for example, authenticating with username/password
to a server as well as performing nickserv authentication after login.

The existing 'auth_method' param remains, and if specified will override
both 'server_auth_method' and 'nick_auth_method'.