NEWS: Write 7.0 version header
Starts out basic, but the main point is to add a link to the
(still-in-progress) Sopel 7 migration guide.

We're on track to include over a hundred merged pull requests and closed
issues in Sopel 7, and the list of changes will include some trivial
stuff that doesn't really affect users or developers much. That's why
the important stuff also gets an entry in the migration guide.