plugins: use a function to get usable plugins
The base function `enumerate_plugins` lists all possible plugins, but:

* Sopel must not load overridden plugins
* Sopel must not configure overriden plugins

so instead, the core uses this new function `get_usable_plugins`, that
makes sure to provide only plugins that can be used, unique per name,
following the right order of precedence:

* extra directories
* homedir directory
* sopel_modules.*
* sopel.modules.*

and the first to be found override the last, meaning that a user can
override core plugins at will.

Note that sopel.coretasks can not and won't be overloaded.