ip: Google changed the canonical hostname of their DNS
dns.google is quite a catchy name, but our tests were failing because of
this sudden (not that I know where they would announce it) change.

On the plus side, we now have robust tests that should correctly handle
both the old and new hostnames, and probably most potential future
changes as well. And they're *actually* using regex syntax, unlike
before with the unescaped brackets and pipes.