reapply fix: interrupt during disconnected phase (9dd772d)
Because of the `time.sleep` during the disconnect phase, the signal
handling is thrown off and a new bot is spawned before the `bot.hasquit`
flag can be checked. Add a check at the top of the loop to see if the
`hasquit` flag was set during a disconnect.

Fixes #1478. Reapplied because of changes from #1493.