core: perform commands on connect
Adds a setting in `core` called `commands_on_connect` based on the updated
`ListAttribute` from #1460. This setting stores a comma separated list
of raw IRC commands (`\`-escaped commas in commands) to execute upon
successful connection to the server. As @dgw put it, "Think ZNC's
perform module, but without the ability to add/remove/rearrange lines
from an IRC query" in #1455.

The commands in the `commands_on_connect` list are executed at the end of
the `startup` procedure. They can also be called by a bot admin with the
`.execute` command.

Note: two `TODO`s were added to adjust docstrings and docs once #1628 is
accepted, since it will change the `ListAttribute` delimiter from
commas to newlines.

Closes #1455