fix: regression in testing individual modules
Using `sopel.tools.target...` in `sopel/test_tools.py` an error because
`sopel.tools` does not seem to be aware of `target`. I expect this is
because `target.py` requires a class from `sopel.tools` and this would
leads to circular imports. Need to `import sopel.tools.target`
separately to avoid this.

`pytest` tests work fine because it knows of `sopel.tools.target` by
that name already from traversing the `rootdir`.