coretasks: fix malformed MODE command in startup()
bot.write() joins the elements of a sequence type (tuple, here) with a
space already, so including a space in `'MODE '` caused Sopel to send
`MODE  SopelsNick +modes`. Technically, that's invalid.

Discovered while trying to test #1470 on a network with actual support
for `echo-message`, when darwin.network disconnected the bot during
registration with the message "ERROR :Received malformed line". Annoying
because I was trying to test something else, but useful in the end.