fix: interrupt during connection phase (before connection accepted)
Before the connection is accepted (and the socket exists), `^C` will
start closing nicely, until `irc.Bot.handle_close`. `handle_close()`
tries to close the socket, even if it may not exist, which leads to a
messy exit.

All that is needed is a quick check to ensure a socket exists before
trying to close it.